Call the health endpoint on login page
We hit the health endpoint when the login page is requested to help kick-start the Serverless RDS instance before it is required for the login process.
Sadly, Serverless v1 can take over the 30s API-GW threshold to start, the health endpoint hits the DB so this should start it up just before.